refactor: simplify MiniMax support to use existing OpenAI path with generic think-tag stripping
Removed the separate minimax_sdk.py, MiniMax provider enum, and dedicated
client singleton.  MiniMax is OpenAI-compatible, so users now configure it
via LLM_SDK=openai + LLM_BASE_URL=https://api.minimax.io/v1.

Added a generic _strip_think_tags() helper in openai_complete that benefits
all reasoning models (MiniMax, DeepSeek, QwQ, etc.) by stripping
<think>...</think> blocks from responses.
