+++ docs/content/docs/(guides)/settings/core.mdx
@@ -18,7 +18,7 @@ Custom base URL for LLM API endpoints. Leave unset to use the provider's default
 </ParamField>
 
 <ParamField path="LLM_SDK" type="string" default="openai">
-LLM provider to use. Supported values: `openai`, `anthropic`, `minimax`
+LLM provider to use. Supported values: `openai`, `anthropic`
 </ParamField>
 
 <ParamField path="LLM_SIMPLE_MODEL" type="string" default="gpt-4.1">
@@ -90,9 +90,10 @@ BLOCK_EMBEDDING_API_KEY=sk-your-openai-key-for-embeddings
 ```
 
 ```bash title="MiniMax Setup"
-# Using MiniMax (OpenAI-compatible API)
+# MiniMax is OpenAI-compatible — use LLM_SDK=openai with a custom base URL
 LLM_API_KEY=your-minimax-api-key
-LLM_SDK=minimax
+LLM_SDK=openai
+LLM_BASE_URL=https://api.minimax.io/v1
 LLM_SIMPLE_MODEL=MiniMax-M2.7
 
 # Keep OpenAI for embeddings (recommended)
