+++ src/server/core/acontext_core/llm/complete/__init__.py
@@ -1,7 +1,6 @@
 from typing import Callable, Awaitable, Mapping, Optional
 from .openai_sdk import openai_complete
 from .anthropic_sdk import anthropic_complete
-from .minimax_sdk import minimax_complete
 from .mock_sdk import mock_complete
 from ...schema.llm import LLMResponse
 from ...schema.result import Result
@@ -14,7 +13,6 @@ COMPLETE_FUNC = Callable[..., Awaitable[LLMResponse]]
 FACTORIES: Mapping[str, COMPLETE_FUNC] = {
     "openai": openai_complete,
     "anthropic": anthropic_complete,
-    "minimax": minimax_complete,
     "mock": mock_complete,
 }
 
@@ -63,7 +61,7 @@ async def llm_sanity_check():
 
 
 def response_to_sendable_message(message: LLMResponse) -> dict:
-    if DEFAULT_CORE_CONFIG.llm_sdk in ("openai", "minimax"):
+    if DEFAULT_CORE_CONFIG.llm_sdk == "openai":
         return message.raw_response.choices[0].message.model_dump()
     elif DEFAULT_CORE_CONFIG.llm_sdk == "anthropic":
         dp = {"role": message.role, "content": []}
